All pages Automation code, resources , config and Allure Setup

+++ tests/test_delete.py
@@ -1,3 +1,4 @@
+import allure
 import pytest
 from selenium.webdriver.common.by import By
 from pages.login_page import LoginPage
@@ -6,7 +7,8 @@ from pages.album_page import AlbumPage
 from pages.delete_page import DeletePage
 
 
-@pytest.mark.usefixtures("driver")   # if you are using pytest fixture for driver
+@pytest.mark.usefixtures("driver") 
+@pytest.mark.order(5)
 def test_delete_flow(driver):
     driver.maximize_window()
     driver.get("https://dev.events.snapdme.com/")
@@ -23,18 +25,26 @@ def test_delete_flow(driver):
 
     # Step 3: Open Album
     album_page = AlbumPage(driver)
-    album_page.open_first_album()
+    create_album_btn = (By.XPATH, "//img[@alt='Create']")
+    if create_album_btn:
+        print("No Albums present. Cannot delete photos.")
+    else:
+        album_page.open_first_album()
 
     # Step 4: Delete Photo
-    delete_page = DeletePage(driver)
-    delete_page.delete_checkbox()
-    delete_page.delete_all()
-    delete_page.confirm_delete()
+        delete_page = DeletePage(driver)
+        delete_buttons = driver.find_elements(By.XPATH, "//button[@class='NoPhoto_chooseFilesButton__hzQAi']")
+        if not delete_buttons:
+            print("No photos present. Cannot delete photos.")
+        else:
+            delete_page.delete_checkbox()
+            delete_page.delete_all()
+            delete_page.confirm_delete()
 
     # Step 5: Delete Album
-    album_msg = delete_page.delete_album()
+    # album_msg = delete_page.delete_album()
     # assert "Deleted" in album_msg.text
 
     # Step 6: Delete Event
-    event_msg = delete_page.delete_event()
+    # event_msg = delete_page.delete_event()
     # assert "Deleted" in event_msg.text
